backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
praw
PyGithub
openai
google-generativeai
